REF: Rename service_backend.py to service_backends.py
